refactor(kernel): phase 3e/1 — extract agent_runtime cluster
Move runtime-control methods (cost/stop/list/suspend/resume/compact/
context_report/kill + watchers) from the giant `impl LibreFangKernel`
into `kernel/agent_runtime.rs`. mod.rs drops from 6 385 to 5 825
(-560).

Methods: session_usage_cost, stop_agent_run, stop_session_run,
list_running_sessions, agent_has_active_session, running_session_ids,
suspend_agent, resume_agent, persist_agent_enabled,
compact_agent_session, compact_agent_session_with_id, context_report,
register_agent_watcher, abort_agent_watchers, kill_agent,
kill_agent_with_purge.

Verified: cargo check / clippy -- -D warnings / cargo check --tests
all green.
